requirements.txt
amqp==5.1.1
anyio==3.7.1
apiclient==1.0.4
billiard==4.1.0
blinker==1.6.2
boto3==1.28.10
botocore==1.31.10
cachetools==5.3.1
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.2
dnspython==2.4.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
google-api-core==2.11.1
google-api-python-client==2.106.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.1
MarkupSafe==2.1.3
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.0
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
shotstack-sdk==0.2.6
six==1.16.0
sniffio==1.3.0
stripe==7.6.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
WTForms==3.0.1
